DEV-603 add authorization checks for delete exp

+++ service-experiment/src/main/java/sg/ncl/service/experiment/logic/ExperimentServiceImpl.java
@@ -2,9 +2,13 @@ package sg.ncl.service.experiment.logic;
 
 import lombok.extern.slf4j.Slf4j;
 import org.json.JSONObject;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sg.ncl.adapter.deterlab.AdapterDeterLab;
+import sg.ncl.common.authentication.Role;
+import sg.ncl.common.exception.base.ForbiddenException;
 import sg.ncl.service.experiment.ExperimentConnectionProperties;
 import sg.ncl.service.experiment.data.jpa.ExperimentEntity;
 import sg.ncl.service.experiment.data.jpa.ExperimentRepository;
@@ -21,6 +25,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
+import java.util.Collection;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -218,6 +223,7 @@ public class ExperimentServiceImpl implements ExperimentService {
      * @param teamName the team where the experiment is in (required by Deterlab so that we delete the correct experiment)
      * @return the deleted experiment object
      * @implNote delete the realization object first, follow by the experiment object (the reverse process of create)
+     * @throws ForbiddenException if user is not the experiment creator and user is not an admin
      */
     @Transactional
     public Experiment deleteExperiment(final Long id, final String teamName) {
@@ -226,6 +232,16 @@ public class ExperimentServiceImpl implements ExperimentService {
 
         RealizationEntity realizationEntity = realizationService.getByExperimentId(id);
         Long realizationId = realizationEntity.getId();
+
+        String contextUserId = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Collection<? extends GrantedAuthority> roleList = SecurityContextHolder.getContext().getAuthentication().getAuthorities();
+
+        log.info("Context user id: {}, Context role: {}", contextUserId, roleList);
+
+        if (!contextUserId.equals(realizationEntity.getUserId()) && !roleList.contains(Role.ADMIN)) {
+            log.warn("Access denied for delete experiment: /{}/ ", id);
+            throw new ForbiddenException("Access denied for delete experiment: expid " + id);
+        }
 
         if (realizationId != null && realizationId > 0) {
             realizationService.deleteRealization(realizationId);
